downgraded Gradle version
This pull request resolves a bug in the application caused by an unexpected upgrade to Android Gradle Plugin (AGP) and Gradle versions. Here's an insight into what was attempted to be accomplished with these changes:

Issue Description:
The bug occurred because the AGP upgrade assistant automatically upgraded Gradle to version 8.x.x, but for the assignments related to the project, the code is intended to use AGP 4.1.2, Gradle version 7.2, and Amazon Corretto JDK version 11.0.20. The AGP upgrade caused issues in the app, and this pull request addresses those problems.

Changes Made:

    Build.gradle and gradle-wrapper.properties:
        Inside the build.gradle file and the gradle-wrapper.properties file, specific values were changed to ensure compatibility with AGP 4.1.2 and Gradle version 7.2. The exact changes can be reviewed in the commit's diff.

    Gradle Scripts File:
        The file delimiter in Gradle scripts was changed from .kts to none. This adjustment aligns with the requirements of the new Gradle version. This change can also be seen in the commit's diff.

    AndroidManifest.xml:
        Modifications were made in the AndroidManifest.xml file to remove specific lines of code that were causing issues. These changes are also visible in the commit's diff.

Recommendation:
I would advise against running the AGP upgrade assistant to Gradle 8.x.x and suggests closing the upgrade pop-up when it appears to avoid potential issues in the future.

The changes in this pull request aim to revert to the desired configuration for the project, ensuring compatibility with AGP 4.1.2, Gradle 7.2, and Amazon Corretto JDK version 11.0.20. If there are any further questions or clarifications needed regarding these changes, please let me know via this pull request or through discord